+++ arcane/src/arcane/hdf5/srcs.cmake
@@ -1,0 +1,4 @@
+set(ARCANE_SOURCES
+  Hdf5Utils.cc
+  Hdf5Utils.h
+)
